requirements.txt
# === Core Utilities ===
pandas>=2.0.0
tqdm>=4.66.0
numpy>=1.25.0

# === Web Scraping ===
requests>=2.31.0
beautifulsoup4>=4.12.0

# === Embeddings + Vector Search ===
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0

# === LangChain Ecosystem (for hybrid / OpenAI use) ===
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# === Local / Offline Models ===
transformers>=4.43.0
torch>=2.3.0
accelerate>=0.31.0

# === Web Interface ===
gradio>=4.44.0

# === Local LLM Runtime ===
ollama>=0.1.0